requirements.txt
annotated-types==0.6.0
anyio==3.7.1
blinker==1.7.0
boto3==1.28.82
botocore==1.31.82
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
distro==1.8.0
dnspython==2.4.2
Flask==3.0.0
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
numpy==1.26.1
openai==1.2.0
Pillow==10.1.0
pydantic==2.4.2
pydantic_core==2.10.1
pymongo==4.6.0
pymongo-auth-aws==1.1.0
PyMuPDF==1.23.6
PyMuPDFb==1.23.6
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
s3transfer==0.7.0
six==1.16.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==3.0.1
